+++ ubo_app/__init__.py
@@ -1,4 +1,10 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107
+import importlib
+import importlib.util
+import os
+import sys
+from pathlib import Path
+
 from kivy.clock import Clock
 
 from ubo_app.menu import MenuApp
@@ -7,6 +13,8 @@ from ubo_app.store.status_icons import (
     IconRegistrationAction,
     IconRegistrationActionPayload,
 )
+
+ROOT_PATH = Path(__file__).parent
 
 
 def main() -> None:
@@ -17,7 +25,7 @@ def main() -> None:
         lambda *_: dispatch(
             [
                 IconRegistrationAction(
-                    type='REGISTER_ICON',
+                    type='STATUS_ICONS_REGISTER',
                     payload=IconRegistrationActionPayload(icon='wifi', priority=-1),
                 )
                 for _ in range(2)
@@ -25,6 +33,28 @@ def main() -> None:
         ),
         3,
     )
+
+    for services_directory_path in [
+        ROOT_PATH.joinpath('services').as_posix(),
+        *(
+            os.environ.get('UBO_SERVICES_PATH', '').split(':')
+            if os.environ.get('UBO_SERVICES_PATH')
+            else []
+        ),
+    ]:
+        if Path(services_directory_path).is_dir():
+            for service_path in Path(services_directory_path).iterdir():
+                name = service_path.name
+                spec = importlib.util.spec_from_file_location(
+                    name,
+                    location=service_path.joinpath('__init__.py').as_posix(),
+                )
+                if not spec:
+                    continue
+                module = importlib.util.module_from_spec(spec)
+                sys.modules[spec.name] = module
+                if spec.loader:
+                    spec.loader.exec_module(module)
 
     app.run()