+++ ubo_app/services/0-keypad/__init__.py
@@ -1,0 +1,6 @@
+# ruff: noqa: D104, N999
+from pathlib import Path
+
+IS_RPI = Path('/etc/rpi-issue').exists()
+if IS_RPI:
+    from . import rpi_keypad