+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+## Version 0.2.1
+
+- feat: loading services dynamically from 'services' directory and direcotires specified in `UBO_SERVICES_PATH`
+- feat: implement keyboard and keypad services
+- feat: implement main store handling main menu events
+
 ## Version 0.1.1
 
-refactor: make it pip-friendly
+- refactor: make it pip-friendly
 
 ## Version 0.1.0
 
-feat: implement a simple version of redux
+- feat: implement a simple version of redux